django-rest-official-tutorial/requirements.txt
django
djangorestframework
pygments
httpie
flake8
coreapi
pyyaml
coreapi-cli
